+++ CHANGES.md
@@ -10,7 +10,7 @@
     cast `Layout` back into a concrete type and extract current layout state
     from it.
 
-  * Export constructor for `Choose` and `LR` from `Module.Layout` to allow
+  * Export constructor for `Choose` and `CLR` from `Module.Layout` to allow
     pattern-matching on the left and right sub-layouts of `Choose l r a`.
 
 ## 0.15 (September 30, 2018)